+++ src/controllers/first-order-feedforward.cpp
@@ -1,5 +1,5 @@
 #include "controllers/first-order-feedforward.hpp"
 
-double FirstOrderFeedforward::compute(const FirstOrderFeedforwardState& setpoint) {
+float FirstOrderFeedforward::compute(const FirstOrderFeedforwardState& setpoint) {
     return constants.Ka * setpoint.acceleration + constants.Kv * setpoint.velocity + constants.Ks * sign(setpoint.velocity);
 }
